requirements.txt
fastapi
uvicorn
langchain-community
transformers
torch
python-dotenv
google-generativeai
